Fixed: The in-game help screen listed CTRL+P instead of PAUSE as pause key
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@8843 01821946-851b-0410-b7f4-f1477c70e57a

+++ io/args.cpp
@@ -109,7 +109,7 @@ while (!cfP->EoF ()) {
 			break;
 		if (pszLine) {
 			int l;
-			for (l = strlen (pszLine); l > 0; l--) {
+			for (l = (int) strlen (pszLine); l > 0; l--) {
 				if (pszLine [l - 1] != '\r')
 					break;
 				}
